% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuranMale.R
\docType{data}
\name{FuranMale}
\alias{FuranMale}
\title{FT-Raman spectra of furan maleimide based self-healing polymer}
\format{A matrix containing 6 spectra by rows with 145 wavenumbers by columns.}
\source{
R. Geitner, J. Koetteritzsch, M. Siegmann, T. Bocklitz, M. Hager,
    U. S. Schubert, S. Graefe, B. Dietzek, M. Schmitt and J. Popp,
    \emph{Phys Chem. Chem. Phys.}, 2015, \strong{17}, 22587-22595.
}
\description{
Six preprocessed FT-Raman spectra of a self-healing polymer.
    The wavenumber region shows the C=C vibrations of furan, maleimide and
    thier respective Diels-Alder adduct. The row names show the measurement
    temperature in degree Celsius, while the column names show the relative
    wavenumber.
}

