% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr2d.R
\name{is.corr2d}
\alias{is.corr2d}
\title{Check for object class "corr2d"}
\usage{
is.corr2d(x)
}
\arguments{
\item{x}{An object which should be check if it is of class "corr2d".}
}
\value{
A logical scalar
}
\description{
The function checks if an object is of class "corr2d".
}
\details{
The function uses the \code{\link{inherits}} function.
}
\examples{
    data(FuranMale, package = "corr2D")
    twod <- corr2d(FuranMale, Ref1 = FuranMale[1, ], corenumber = 1)
    
    # TRUE
    is.corr2d(twod) 
    # FALSE
    is.corr2d(2) 

}
