% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr2d_plot.R
\name{plot_corr2d}
\alias{plot_corr2d}
\title{Plot two-dimensional correlation spectra.}
\usage{
plot_corr2d(Obj, what = Re(Obj$FT), specx = Obj$Ref1, specy = Obj$Ref2,
  xlim = NULL, ylim = NULL, xlab = expression(nu[1]),
  ylab = expression(nu[2]), Contour = TRUE, axes = 3, Legend = TRUE,
  N = 20, zlim = NULL, Cutout = NULL, ...)
}
\arguments{
\item{Obj}{List from \code{corr2d} containing the 2D correlation data.}

\item{what}{Real numeric matrix containing the z-values that should be plotted.}

\item{specx, specy}{Numeric vector containing the data that should be plotted
on top (\code{specx}) and/or on the left (\code{specy}) of
the 2D spectrum. \code{Mat}, \code{specx} and/or \code{specy} should
have the same dimensions, respectively. If \code{NULL} nothing will
be plotted.}

\item{xlim, ylim}{Numeric vector with two values indicating the borders
of the 2D plot. Also truncates \code{specx} and/or \code{specy} to
match the new plot range.}

\item{xlab, ylab}{Character or expression containing the text that will
be plotted on the bottom (\code{xlab}) and/or to the right
(\code{ylab}) of the 2D plot. Labels can be suppressed with \code{NA}.}

\item{Contour}{Logical: Should a contour (\code{TRUE}) or image
(\code{FALSE}) be drawn?}

\item{axes}{Integer ranging from 0 to 3. Should the axis of the 2D plot
be drawn? "0" means no axes, "1" only bottom axis, "2" only right axis and
"3" both axes are drawn.}

\item{Legend}{Logical: Should a color legend be plotted in the top
right corner?}

\item{N}{Positive, non-zero integer indicating how many contour or image
levels should be plotted.}

\item{zlim}{Numeric vector with two values defining the z-range of the 2D
plot.}

\item{Cutout}{Numeric vector with two values defining which z-values should
not be plotted. Use with care, because this can generate misleading
2D plots.}

\item{...}{Additional arguments either passed to
\code{\link[graphics]{image}} or \code{\link[graphics]{contour}}. Can
include graphics parameters \code{\link[graphics]{par}} which are in
part also used by other functions. This includes \code{lwd} (sets line
width for the 1D plots on the sides, the central 2D contour plot, the
axes and the surronding box), \code{cex} (influences axes and thier
labels) and \code{cex.axis} (influences legend labels).}
}
\description{
\code{plot_corr2d} plots two-dimensional correlation spectra either
    as an image or a contour plot. Red color indicates positive
    correlations, while blue color shows negative ones.
}
\details{
For the synchronous correlation spectrum the real component (\code{Re})
    of the complex correlation matrix must be plotted. The asynchronous
    spectrum is the respective imaginary component (\code{Im}).
    \code{Cutout} can be used to leave out smaller (noise) contributions,
    but should be used with care as it can be used to create misleading
    2D correlation plots. See references for interpretation rules (so
    called Noda rules).
}
\examples{
    data(FuranMale, package = "corr2D")
    twod <- corr2d(FuranMale, Ref1 = FuranMale[1, ], corenumber = 1)
    
    plot_corr2d(twod, xlab = expression(paste("relative Wavenumber" / cm^-1)),
                      ylab = expression(paste("relative Wavenumber" / cm^-1)))

}
\references{
For interpretation rules see:
    I. Noda (2006) <DOI:10.1016/j.molstruc.2005.12.060>
}
\seealso{
See \code{\link{plot_corr2din3d}} for 3D plots.
}
